requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.5.0
motor==3.3.1
email-validator==1.3.1
python-dateutil==2.8.2
pydantic==1.10.7
python-slugify==8.0.1
gunicorn==20.1.0
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0
